+++ app/src/main/java/com/robertconstantindinescu/woutapp/core/util/Paginator.kt
@@ -2,4 +2,5 @@ package com.robertconstantindinescu.woutapp.core.util
 
 interface Paginator {
     suspend fun loadNextPosts()
+    fun reset()
 }
