+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_authentication/presentation/register/SignUpScreen.kt
@@ -13,10 +13,7 @@ import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.material.ScaffoldState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
-import androidx.compose.material3.Button
-import androidx.compose.material3.Icon
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
+import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
@@ -44,6 +41,7 @@ import com.robertconstantindinescu.woutapp.core.util.UiText
 import com.robertconstantindinescu.woutapp.feature_authentication.domain.util.AuthError
 import kotlinx.coroutines.flow.collectLatest
 
+@ExperimentalMaterial3Api
 @Composable
 fun SignUpScreen(
     scaffoldState: ScaffoldState,
