+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_create_post/presentation/CreatePostScreen.kt
@@ -4,6 +4,7 @@ import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
@@ -26,6 +27,7 @@ import com.robertconstantindinescu.woutapp.feature_create_post.presentation.util
 import com.robertconstantindinescu.woutapp.feature_create_post.presentation.util.CreatePostConstants.PAGING_INDICATOR_WIDTH
 import kotlinx.coroutines.flow.collectLatest
 
+@ExperimentalMaterial3Api
 @ExperimentalPagerApi
 @Composable
 fun CreatePostScreen(
