+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_get_posts/domain/use_case/PostUseCases.kt
@@ -1,0 +1,6 @@
+package com.robertconstantindinescu.woutapp.feature_get_posts.domain.use_case
+
+data class PostUseCases(
+    val mainFeedGetAllPostsUseCase: MainFeedGetAllPostsUseCase,
+    val personalGetAllPostsUseCase: PersonalGetAllPostsUseCase
+)
