+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_get_posts/presentation/main_feed_screen/MainFeedEvent.kt
@@ -1,0 +1,7 @@
+package com.robertconstantindinescu.woutapp.feature_get_posts.presentation.main_feed_screen
+
+sealed class MainFeedEvent {
+    //Add to favorites
+    //Share
+    //Suscribe to evenrt
+}
