+++ app/src/main/java/com/robertconstantindinescu/woutapp/feature_get_posts/util/MainFeedConstants.kt
@@ -1,0 +1,5 @@
+package com.robertconstantindinescu.woutapp.feature_get_posts.util
+
+object MainFeedConstants {
+    const val DEFAULT_PAGE_SIZE = 15
+}
