--- app/src/main/java/com/robertconstantindinescu/woutapp/feature_main_feed/domain/use_case/MainFeedUseCases.kt
@@ -1,5 +1,0 @@
-package com.robertconstantindinescu.woutapp.feature_main_feed.domain.use_case
-
-data class MainFeedUseCases(
-    val getAllPostsUseCase: GetAllPostsUseCase
-)
