--- app/src/main/java/com/robertconstantindinescu/woutapp/feature_main_feed/presentation/MainFeedEvent.kt
@@ -1,7 +1,0 @@
-package com.robertconstantindinescu.woutapp.feature_main_feed.presentation
-
-sealed class MainFeedEvent {
-    //Add to favorites
-    //Share
-    //Suscribe to evenrt
-}
