--- app/src/main/java/com/robertconstantindinescu/woutapp/feature_main_feed/util/MainFeedConstants.kt
@@ -1,5 +1,0 @@
-package com.robertconstantindinescu.woutapp.feature_main_feed.util
-
-object MainFeedConstants {
-    const val DEFAULT_PAGE_SIZE = 10
-}
